+++ doc/source/changelog.rst
@@ -8,6 +8,7 @@ v1.2.0
 * Implemented `Switch to ruff for linting and formatting. <https://github.com/bckohan/django-typer/issues/56>`_
 * Implemented `Add a wrapper for typer's echo/secho <https://github.com/bckohan/django-typer/issues/55>`_
 * Implemented `Better test organization. <https://github.com/bckohan/django-typer/issues/34>`_
+* Implemented `Add completer/parser for GenericIPAddressField. <https://github.com/bckohan/django-typer/issues/12>`_
 
 
 v1.1.2